* more integer type explicitness [rtj]

+++ src/hddm-cpp.cpp
@@ -2122,7 +2122,7 @@ void CodeBuilder::checkConsistency(DOMElement* el, DOMElement* elref)
 
    DOMNamedNodeMap* oldAttr = elref->getAttributes();
    DOMNamedNodeMap* newAttr = el->getAttributes();
-   unsigned int listLength = oldAttr->getLength();
+   size_t listLength = oldAttr->getLength();
    for (unsigned int n = 0; n < listLength; n++)
    {
       XtString nameS(oldAttr->item(n)->getNodeName());
@@ -2642,7 +2642,7 @@ void CodeBuilder::constructGroup(DOMElement* el)
 
    parentList.push_back(el);
    DOMNodeList* contList = el->getChildNodes();
-   int contLength = contList->getLength();
+   size_t contLength = contList->getLength();
    for (int c = 0; c < contLength; c++)
    {
       DOMNode* cont = contList->item(c);
@@ -4235,7 +4235,7 @@ void CodeBuilder::writeStreamers(DOMElement* el)
 
    std::vector<XtString> contV;
    DOMNodeList* contList = el->getChildNodes();
-   int contListLength = contList->getLength();
+   size_t contListLength = contList->getLength();
    for (int c = 0; c < contListLength; c++)
    {
       DOMNode* node = contList->item(c);
@@ -4511,7 +4511,7 @@ void CodeBuilder::constructDocument(DOMElement* el)
    XtString tagS(el->getTagName());
    hFile << "<" << tagS;
    DOMNamedNodeMap* attrList = el->getAttributes();
-   int attrListLength = attrList->getLength();
+   size_t attrListLength = attrList->getLength();
    for (int a = 0; a < attrListLength; a++)
    {
       DOMNode* node = attrList->item(a);
@@ -4521,7 +4521,7 @@ void CodeBuilder::constructDocument(DOMElement* el)
    }
 
    DOMNodeList* contList = el->getChildNodes();
-   int contListLength = contList->getLength();
+   size_t contListLength = contList->getLength();
    if (contListLength > 0)
    {
       hFile << ">\\n\"" << std::endl;
